Make kernelspecs a single API call
