+++ src/index.ts
@@ -7,5 +7,4 @@ export * from './contents';
 export * from './ikernel';
 export * from './isession';
 export * from './kernel';
-export * from './kernelselector';
 export * from './session';
