+++ src/tsconfig.json
@@ -19,7 +19,6 @@
     "config.ts",
     "contents.ts",
     "kernel.ts",
-    "kernelselector.ts",
     "ikernel.ts",
     "isession.ts",
     "serialize.ts",
