+++ src/validate.ts
@@ -2,7 +2,7 @@
 // Distributed under the terms of the Modified BSD License.
 'use strict';
 
-import { IKernelId } from './ikernel';
+import { IKernelId, IKernelSpecId } from './ikernel';
 
 import { INotebookId, ISessionId } from './isession';
 
@@ -47,4 +47,37 @@ function validateNotebookId(model: INotebookId): void {
    if ((!model.hasOwnProperty('path')) || (typeof model.path !== 'string')) {
      throw Error('Invalid Notebook Model');
    }
+}
+
+
+/**
+ * Validate an object as being of IKernelSpecID type.
+ */
+ export
+function validateKernelSpec(info: IKernelSpecId): void {
+  var err = new Error("Invalid KernelSpec Model");
+  if (!info.hasOwnProperty('name') || typeof info.name !== 'string') {
+    throw err;
+  }
+  if (!info.hasOwnProperty('spec') || !info.hasOwnProperty('resources')) {
+    throw err;
+  }
+  var spec = info.spec;
+  if (!spec.hasOwnProperty('language') || typeof spec.language !== 'string') {
+    throw err;
+  }
+  if (!spec.hasOwnProperty('display_name') ||
+      typeof spec.display_name !== 'string') {
+    throw err;
+  }
+  if (!spec.hasOwnProperty('argv') || !Array.isArray(spec.argv)) {
+    throw err;
+  }
+  if (!spec.hasOwnProperty('codemirror_mode') ||
+      typeof spec.codemirror_mode !== 'string') {
+    throw err;
+  }
+  if (!spec.hasOwnProperty('env') || !spec.hasOwnProperty('help_links')) {
+    throw err;
+  }
 }
