+++ test/tsconfig.json
@@ -20,7 +20,7 @@
     "src/testconfig.ts",
     "src/testcontents.ts",
     "src/testkernel.ts",
-    "src/testkernelselector.ts",
+    "src/testkernelspecs.ts",
     "src/testsession.ts",
     "src/testmocksocket.ts",
     "src/testmockxhr.ts",
